Use thread pool for git commands
