+++ clowder/clowder/clowder_repo.py
@@ -5,13 +5,13 @@ from __future__ import print_function
 import atexit
 import errno
 import os
+import subprocess
 import sys
 
 from termcolor import colored
 
 import clowder.util.formatting as fmt
 from clowder.git.project_repo import ProjectRepo
-from clowder.process_pool import execute_command
 from clowder.util.connectivity import is_offline
 from clowder.util.file_system import remove_directory
 
@@ -125,17 +125,17 @@ class ClowderRepo(object):
     def pull(self):
         """Pull clowder repo upstream changes"""
         repo = ProjectRepo(self.clowder_path, self.remote, self.default_ref)
-        repo.pull(self.clowder_path)
+        repo.pull()
 
     def push(self):
         """Push clowder repo changes"""
         repo = ProjectRepo(self.clowder_path, self.remote, self.default_ref)
-        repo.push(self.clowder_path)
+        repo.push()
 
     def run_command(self, command):
         """Run command in clowder repo"""
         print(fmt.command(command))
-        return_code = execute_command(command.split(), self.clowder_path)
+        return_code = execute_subprocess_command(command.split(), self.clowder_path)
         if return_code != 0:
             print(fmt.command_failed_error(command))
             sys.exit(return_code)
@@ -152,6 +152,30 @@ class ClowderRepo(object):
             ProjectRepo.validation(self.clowder_path)
             print()
             sys.exit(1)
+
+
+def subprocess_exit_handler(process):
+    """terminate subprocess"""
+    try:
+        os.kill(process.pid, 0)
+        process.kill()
+    except:
+        pass
+
+
+def execute_subprocess_command(command, path, shell=True, env=None, stdout=None, stderr=None):
+    """Run subprocess command"""
+    try:
+        process = subprocess.Popen(' '.join(command), shell=shell, env=env, cwd=path,
+                                   stdout=stdout, stderr=stderr)
+        atexit.register(subprocess_exit_handler)
+        process.communicate()
+    except (KeyboardInterrupt, SystemExit):
+        raise
+    except:
+        raise
+    else:
+        return process.returncode
 
 
 def force_symlink(file1, file2):
